Refactor and merge QueryDBState into StorageActorState
All classes implementing one also needs the other.
Separation does not have added value.
