+++ engine/src/main/java/org/cafienne/storage/actormodel/message/StorageActionUpdated.scala
@@ -1,0 +1,5 @@
+package org.cafienne.storage.actormodel.message
+
+trait StorageActionUpdated extends StorageEvent {
+
+}
