+++ engine/src/main/java/org/cafienne/storage/actormodel/state/ProcessState.scala
@@ -20,7 +20,7 @@ package org.cafienne.storage.actormodel.state
 import org.cafienne.storage.actormodel.ActorMetadata
 import org.cafienne.storage.querydb.ProcessStorage
 
-trait ProcessState extends QueryDBState {
+trait ProcessState extends StorageActorState {
   override val dbStorage: ProcessStorage = new ProcessStorage(actor.caseSystem.queryDB.writer)
 
   override def findCascadingChildren(): Seq[ActorMetadata] = Seq()
