+++ engine/src/main/java/org/cafienne/storage/archival/ActorDataArchiver.scala
@@ -17,10 +17,10 @@
 
 package org.cafienne.storage.archival
 
-import org.apache.pekko.persistence.DeleteMessagesSuccess
 import com.typesafe.scalalogging.LazyLogging
+import org.apache.pekko.persistence.DeleteMessagesSuccess
 import org.cafienne.storage.actormodel.message.StorageEvent
-import org.cafienne.storage.actormodel.{ActorMetadata, ActorType, QueryDBStorageActor}
+import org.cafienne.storage.actormodel.{ActorMetadata, ActorType, StorageActor}
 import org.cafienne.storage.archival.command.ArchiveActorData
 import org.cafienne.storage.archival.event._
 import org.cafienne.storage.archival.event.cmmn.ModelActorArchived
@@ -28,7 +28,7 @@ import org.cafienne.storage.archival.response.{ArchivalCompleted, ArchivalReject
 import org.cafienne.storage.archival.state.{ArchivalState, CaseArchivalState, ProcessArchivalState}
 import org.cafienne.system.CaseSystem
 
-class ActorDataArchiver(override val caseSystem: CaseSystem, override val metadata: ActorMetadata) extends QueryDBStorageActor[ArchivalState] with LazyLogging {
+class ActorDataArchiver(override val caseSystem: CaseSystem, override val metadata: ActorMetadata) extends StorageActor[ArchivalState] with LazyLogging {
 
   printLogMessage(s"\n========== Launching Storage Archival Service ${metadata.path}")
